requirements.txt
alembic==1.4.2
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
click==7.1.2
dnspython==1.16.0
email-validator==1.0.5
eventlet==0.25.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gevent==20.5.0
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==20.0.4
idna==2.9
imutils==0.5.3
git+http://github.com/m-housh/io-blueprint.git
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.2
MarkupSafe==1.1.1
monotonic==1.5
numpy==1.18.3
opencv-python==4.2.0.34
Pillow==7.1.2
psycopg2-binary==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.12.1
python-socketio==4.5.1
scipy==1.4.1
six==1.14.0
SQLAlchemy==1.3.16
torch==1.4.0
torchvision==0.5.0
Werkzeug==1.0.1
WTForms==2.3.1
